+++ models/DataObject/ClassDefinition/Data/ManyToManyObjectRelation.php
@@ -721,24 +721,6 @@ class ManyToManyObjectRelation extends AbstractRelations implements QueryResourc
         return parent::addListingFilter($listing, $data, $operator);
     }
 
-    /**
-     * Filter by relation feature
-     *
-     *
-     */
-    public function getFilterConditionExt(mixed $value, string $operator, array $params = []): string
-    {
-        $prefix = '';
-        $name = $params['name'] ?: $this->name;
-
-        if ($params['brickPrefix']) {
-            // The brick prefix is always quoted and with a dot suffix, so removing the first
-            // and second last character to unquote
-            $prefix = substr($params['brickPrefix'], 1, -2) . substr($params['brickPrefix'], -1);
-        }
-
-        return $this->getRelationFilterCondition($value, $operator, $prefix . $name);
-    }
 
     public function getQueryColumnType(): string
     {
