+++ models/DataObject/ClassDefinition/Data/ManyToManyRelation.php
@@ -783,25 +783,6 @@ class ManyToManyRelation extends AbstractRelations implements QueryResourcePersi
         throw new InvalidArgumentException('Filtering '.__CLASS__.' does only support "=" operator');
     }
 
-    /**
-     * Filter by relation feature
-     *
-     *
-     */
-    public function getFilterConditionExt(mixed $value, string $operator, array $params = []): string
-    {
-        $prefix = '';
-        $name = $params['name'] ?: $this->name;
-
-        if ($params['brickPrefix']) {
-            // The brick prefix is always quoted and with a dot suffix, so removing the first
-            // and second last character to unquote
-            $prefix = substr($params['brickPrefix'], 1, -2) . substr($params['brickPrefix'], -1);
-        }
-
-        return $this->getRelationFilterCondition($value, $operator, $prefix . $name);
-    }
-
     public function getQueryColumnType(): string
     {
         return 'text';
