+++ models/DataObject/ClassDefinition/Data/Relations/ManyToManyRelationTrait.php
@@ -16,6 +16,7 @@ declare(strict_types=1);
 
 namespace Pimcore\Model\DataObject\ClassDefinition\Data\Relations;
 
+use Pimcore\Db;
 use Pimcore\Model\DataObject;
 use Pimcore\Model\DataObject\Concrete;
 use Pimcore\Model\DataObject\Fieldcollection\Data\AbstractData;
@@ -80,4 +81,32 @@ trait ManyToManyRelationTrait
 
         return $data;
     }
+
+    /**
+     * Filter by relation feature
+     *
+     *
+     */
+    public function getFilterConditionExt(mixed $value, string $operator, array $params = []): string
+    {
+        $prefix = '';
+        $name = $params['name'] ?: $this->name;
+
+        if ($params['brickPrefix']) {
+            $prefix = $params['brickPrefix'];
+            // The brick prefix is always quoted and with a dot suffix, so removing the first
+            // and second last character to unquote
+            $quoteIdentifierSymbol  = substr(Db::get()->quoteIdentifier(''), 0 ,1);
+            if (
+                substr($prefix, 1, -2) === $quoteIdentifierSymbol &&
+                substr($prefix, -1) === $quoteIdentifierSymbol
+            ) {
+                $prefix = substr($prefix, 1, -2) . substr($prefix, -1);
+            }
+        }
+
+        return $this->getRelationFilterCondition($value, $operator, $prefix . $name);
+    }
+
+
 }
